backend/requirements.txt
# Backend dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Redis & Caching
redis==5.0.1
hiredis==2.3.2

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
bcrypt==4.1.2
cryptography==42.0.0

# Background tasks
celery==5.3.6
flower==2.0.1

# WebSocket
python-socketio==5.11.0
websockets==12.0

# WhatsApp & External APIs
httpx==0.26.0
aiohttp==3.9.1
requests==2.31.0

# OAuth
authlib==1.3.0

# Payment
stripe==7.11.0

# Email
fastapi-mail==1.4.1

# Utilities
python-slugify==8.0.1
phonenumbers==8.13.27
pytz==2023.3
python-dateutil==2.8.2

# File handling
python-magic==0.4.27
pillow==10.2.0

# Validation & serialization
email-validator==2.1.0
pydantic[email]==2.5.3

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0

# Code quality
black==23.12.1
flake8==7.0.0
mypy==1.8.0

# Monitoring (optional)
sentry-sdk==1.39.2
